+++ CHANGELOG.en-US.md
@@ -60,6 +60,8 @@ timeline: true
 - 🐞 `<pro>Select`: Fixed an issue where the `restrict` attribute could not restrict Chinese characters.
 - 🐞 `<pro>PerformanceTable`: Fixed the class name error when using the `rowClassName` property to dynamically set the class name and switch the current row.
 - 🐞 `<pro>FormField`: Fixed an issue where input component `onCompositionStart` and `onCompositionEnd` events did not take effect.
+- 🐞 `<pro>TextField`: Fix to copy error when `navigator.clipboard` does not exist.
+- 🐞 `<pro>DatePicker`: Fixed an issue where text was copied incorrectly.
 
 ## 1.6.6
